+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 3.0.2 Jul 26, 2024
+
+- Implements `onPermissionRequest()` to call a new `eventProcessor.onPermissionRequest(permissionRequest: PermissionRequest)` callback allowing clients to grant or deny permission requests, or request permissions (e.g. camera, record audio). This is sometimes required for checkouts that use features that require verifying identity.
+
 ## 3.0.1 May 31, 2024
 
 - Call `onPause()` on the WebView as it's created if preloading, and `onResume()` when it's presented, so the Page Visibility API reports correct values.